% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationfunctions.R
\name{autoCorrelationPlot}
\alias{autoCorrelationPlot}
\title{Plot a correlation matrix}
\usage{
autoCorrelationPlot(df, use = "everything", method = c("pearson", "kendall",
  "spearman"), cluster = FALSE, interactiveplot = FALSE)
}
\arguments{
\item{df}{Dataframe to calculate multivariate correlations from}

\item{use}{Same as \code{use} in the \code{cor} function.}

\item{method}{Same as \code{method} in the \code{cor} function.}

\item{cluster}{If \code{cluster = TRUE} then the correlation matrix will be
reordered using the \code{"hclust"} method in \code{corrplot::corrMatOrder}}

\item{interactiveplot}{If \code{FALSE}, the default, returns a \code{corrplot}
visualisation of the correlation matrix. If \code{TRUE}, returns a interactive
\code{heatmaply} visualisation of the correlation matrix}
}
\value{
Correlation plot of the provided correlation matrix
}
\description{
Given a dataframe this function will generate a
correlation plot. Uses \code{corrplot::corrplot} for non-interactive plotting
and \code{heatmaply::heatmaply} for interactive plotting
}
