% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEvisualisationfunctions.R
\name{autoDensityPlot_}
\alias{autoDensityPlot_}
\title{Create a densiy estimate of a target variable}
\usage{
autoDensityPlot_(df, target, interactiveplot = FALSE, xlabel = NULL, ...)
}
\arguments{
\item{df}{Dataframe that contains the target variable}

\item{target}{String giving the name of the target variable}

\item{interactiveplot}{If \code{FALSE}, the default, returns a \code{ggplot2}
visualisation of the density estimate of the target variable. If
\code{TRUE}, returns an interactive \code{plotly} visualisation of the histogram.}

\item{xlabel}{Provide a character to override the default label for the
x axis}

\item{...}{Other arguments passed onto \code{geom_density}, such as
\code{colour = "red"} or \code{size = 2}}
}
\value{
A density estimate visualisation of the target variable from the
provided data
}
\description{
Standard evaluation version of autoDensityPlot. Generates either
 a static \code{ggplot} or interactive \code{plotly} density estimate
 visualisation of the chosen target variable using \code{geom_density}.
}
