% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEvisualisationfunctions.R
\name{autoHistogramPlot_}
\alias{autoHistogramPlot_}
\title{Create a histogram of a target variable}
\usage{
autoHistogramPlot_(df, target, binwidth = NULL, interactiveplot = FALSE,
  xlabel = NULL, ..., stat = "bin")
}
\arguments{
\item{df}{Dataframe that contains the target variable}

\item{target}{String giving the name of the target variable}

\item{binwidth}{The width of the bins. The default \code{NULL} gives
standard \code{geom_histogram} behaviour, but should be overridden with
your own value}

\item{interactiveplot}{If \code{FALSE}, the default, returns a ggplot
visualisation of the histogram of the target variable. If \code{TRUE},
returns an interactive plotly visualisation of the histogram.}

\item{xlabel}{Provide a character to override the default label for the
x axis}

\item{...}{Other arguments passed onto \code{geom_histogram}, such as
\code{colour = "blue"} or \code{fill = NA}}

\item{stat}{Defaults to \code{stat = "bin"} for numeric data. If categorical data
is passed then the function will automatically change to \code{stat = "count"}}
}
\value{
A histogram of the target variable from the provided data
}
\description{
Standard evaluation version of autoHistogramPlot.
Generates either a static \code{ggplot} or interactive \code{plotly}
histogram visualisation of the chosen target variable.
}
