% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenfunctions.R
\name{autoMarkdown}
\alias{autoMarkdown}
\title{Automated generation of R Markdown documents from R Scripts}
\usage{
autoMarkdown(filename, rmdfile = NULL, quiet = TRUE, render = FALSE,
  overwrite = FALSE, divider = "#.#")
}
\arguments{
\item{filename}{Either a single, or vector of, R script filenames in the order which you wish them
to be written to R Markdown}

\item{rmdfile}{The name of the rmdfile to write to, with the \code{.Rmd} extension}

\item{quiet}{The default is \code{quiet = TRUE} which creates 'quiet' chunks. These code chunk will display
the results of the code only}

\item{render}{Automatically render the resulting R Markdown document}

\item{overwrite}{If \code{overwrite = TRUE} the provided \code{.Rmd} file will be completely overwritten}

\item{divider}{A unique character used to identify where the user wishes to split their script into separate
code chunks and comment sections}
}
\description{
Quickly and easily generate R Markdown documents from an R script.
Combine any number of R scripts into a single R Markdown document.
}
