% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenfunctions.R
\name{createDocument}
\alias{createDocument}
\title{Auto create R Markdown documents}
\usage{
createDocument(df, target, template, reporttitle, reportauthor, documentname,
  documentwd, writedoc = TRUE)
}
\arguments{
\item{df}{Dataframe containing the variables of interest}

\item{target}{For univariate reporting - the target variable}

\item{template}{The name of the template document to use. Currently only \code{"univariate"}
is available}

\item{reporttitle}{Character giving the report title in the YAML header}

\item{reportauthor}{Character giving the report author's name in the YAML
header}

\item{documentname}{Character giving the filename of the generated report}

\item{documentwd}{Character giving the directory to which the report should
be written}

\item{writedoc}{Default \code{writedoc = TRUE} will auto generate the final document.
Changing to \code{writedoc = FALSE} will prevent the input .Rmd from being rendered to the
specified input (currently only .html) to allow review}
}
\value{
If \code{writedoc = TRUE} writes .Rmd file and auto generated document to \code{documentwd}.
If \code{writedoc = FALSE} writes .Rmd file only.
}
\description{
Automatically generate data exploration reports using
R Markdown templates.
}
