% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowritefunctions.R
\name{createFile}
\alias{createFile}
\title{Generate a named R Markdown file in a given directory}
\usage{
createFile(filename, filedir = "current")
}
\arguments{
\item{filename}{Character giving the desired name of the R Markdown file, does
not require the \code{.Rmd} extension}

\item{filedir}{Default \code{"current"} will write the file to the current working
directory, given by \code{getwd()}. Provide a character to give an alternative}
}
\value{
Returns the name of the file, including the \code{.Rmd} extension so that
insert functions can be chained after
}
\description{
Generate a named R Markdown file in a given directory
}
