% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationfunctions.R
\name{ggdensity}
\alias{ggdensity}
\title{Create a densiy estimate of a target variable}
\usage{
ggdensity(df, target, interactiveplot = FALSE)
}
\arguments{
\item{df}{Dataframe that contains the target variable}

\item{target}{String giving the name of the target variable}

\item{interactiveplot}{If \code{FALSE}, the default, returns a ggplot
visualisation of the density estimate of the target variable. If
\code{TRUE}, returns a interactive plotly visualisation of the histogram.}
}
\description{
Generates either a static ggplot or interactive plotly
density estimate visualisation of the chosen target variable.
}
