% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationfunctions.R
\name{gghistogram}
\alias{gghistogram}
\title{Create a histogram of a target variable}
\usage{
gghistogram(df, target, interactiveplot = FALSE)
}
\arguments{
\item{df}{Dataframe that contains the target variable}

\item{target}{String giving the name of the target variable}

\item{interactiveplot}{If \code{FALSE}, the default, returns a ggplot2
visualisation of the histogram of the target variable. If \code{TRUE},
returns a interactive plotly visualisation of the histogram.}
}
\description{
Generates either a static ggplot2 or interactive plotly
histogram visualisation of the chosen target variable.
}
