% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowritefunctions.R
\name{insertChunk}
\alias{insertChunk}
\title{Insert a code chunk into an R Markdown file}
\usage{
insertChunk(filename, object, title = NULL)
}
\arguments{
\item{filename}{Character giving the name of the R Markdown file that the
chunk should be inserted into, minus the \code{.Rmd} extension}

\item{object}{Can be any unquoted line of R code, i.e. a call to a function}

\item{title}{Character giving an optional markdown header that will can
included before the code chunk e.g. of the form \code{# Header, ## Header,
### Header}}
}
\value{
Returns \code{filename} so that functions can be chained to write
to the same file consecutively
}
\description{
This function can be used to insert any line of code
into an R Markdown file. The chunk will be inserted into the file called
\code{filename}, which must exist in the current working directory. Chaining
with \code{\%>\%} is possible, by returning \code{filename} from the function.
}
