% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkfunctions.R
\name{insertQuietChunk}
\alias{insertQuietChunk}
\title{Insert a 'quiet' code chunk into an R Markdown file}
\usage{
insertQuietChunk(filename, object, chunktitle = NULL)
}
\arguments{
\item{filename}{Character giving the name of the R Markdown file that the
chunk should be inserted into, minus the \code{.Rmd} extension}

\item{object}{Can be any quoted or unquoted line of R code, i.e. a call to a function to insert a
single line. To insert more than one line, pass as a vector of type \code{"expression"}.}

\item{chunktitle}{Single or vector of character/s giving an optional markdown header that will be
included before the code chunk e.g. of the form \code{"# Header", "## Header",
"### Header"}. Pass as a vector of titles of equal length to object expressions
if multiple headers required.}
}
\value{
Returns \code{filename} so that functions can be chained to write
to the same file consecutively
}
\description{
This function can be used to insert any line or lines of code
into an R Markdown file 'quietly'. The chunk will be inserted into the file called
\code{filename.Rmd}, which must exist in the current working directory. Chaining
with \code{\%>\%} is possible, by returning \code{filename} from the function.
\code{insertQuietChunk} will insert a code chunk that will only show the results
of the code in the rendered document.
}
\examples{

\dontrun{

## Insert single chunk, chained from createFile() function

createFile("Mtcars") \%>\% insertYAML(title = "Test", author = "J. Bloggs") \%>\%
                         insertLibraries() \%>\%
                         insertQuietChunk(summary(mtcars), title = "#Summary")

                         insertChunk("summary(mtcars)", title = "#Summary") would also work

## Insert multiple chunks to existing file, with headers

insertQuietChunk("Mtcars.Rmd",
                 expression(
                      plot(mtcars$cyl, mtcars$hp),
                      multivariateCorrelation(mtcars)),
                            title= c("##Scatter Plot",
                                     "##Correlations"))
}
}
