% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowritefunctions.R
\name{insertYAML}
\alias{insertYAML}
\title{Insert a YAML header into the R Markdown file}
\usage{
insertYAML(filename, title = NULL, author = NULL, date = NULL,
  output = "html_document")
}
\arguments{
\item{filename}{The name of the file, including the .Rmd extension, into
which the YAML header should be inserted}

\item{title}{Character giving the title of the document}

\item{author}{Character giving the author's name}

\item{date}{If \code{NULL} inserts \code{Sys.Date()} into the date field.}

\item{output}{The document type that will be generated, e.g. \code{"html_document"
"pdf"}}
}
\value{
Returns the filename to allow chaining with other insert functions
}
\description{
Insert a YAML header into the R Markdown file
}
