% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOutlierFunction.R
\name{MultivariateOutlier}
\alias{MultivariateOutlier}
\title{Multivariate outlier detection for numerical variables}
\usage{
MultivariateOutlier(df)
}
\arguments{
\item{df, }{dataframe of numerical variables}
}
\value{
Returns a list of 2 objects (1: the original dataframe with an appended column identifying outliers (outliers==0), 2: A plot of the first two components of PCA, outliers identified in red)
}
\description{
This function will return a list of 2 objects (1: the original dataframe with an appended column identifying outliers (outliers==0), 2: A plot of the first two components of PCA, outliers identified in red)
}
