% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateOutlierFunction.R
\name{OptimalMinPoints}
\alias{OptimalMinPoints}
\title{Calculation of the optimal value for the minimum points parameter used in dbscan}
\usage{
OptimalMinPoints(df)
}
\arguments{
\item{df, }{a dataframe of numerical variables}
}
\value{
This function returns a list of the optimal values for the minimum points and epsilon parameters based on the inputted dataframe.
}
\description{
This function calculates the optimum value for the minimum points parameter in dbscan, based on the optimal epsilon value.
The minimum points parameter specifies how many neighbours a point should have to be included in a cluster.
}
