% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFunctions.R
\name{Outlier}
\alias{Outlier}
\title{Outlier detection and outlier summary statistics}
\usage{
Outlier(x)
}
\arguments{
\item{x}{= variable}
}
\value{
list of 6 objects (1:mild outliers, 2: extreme outliers, 3: mild outlier thresholds, 4: extreme outlier thresholds, 5: Percentage mild and extreme outleirs, 6: variable with and  without mild and extreme outliers)
}
\description{
This function will return a list of 6 objects detailing identified mild and extreme outliers, outlier thresholds and summary statistics for the variable with and without mild/extreme outliers
}
