% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFunctions.R
\name{OutlierPercentage}
\alias{OutlierPercentage}
\title{Percentage of observations which are mild and extreme outliers}
\usage{
<<<<<<< HEAD
OutlierPercentage(outliers, x)
}
\arguments{
\item{outliers, }{The list output from the OutlierValues function, x, the variable of interest}
=======
OutlierPercentage(x)
}
\arguments{
\item{x}{= variable}
>>>>>>> master
}
\value{
A data frame with the percentage of observations which are mild and extreme outliers
}
\description{
This function will calculate what percentage of observations are mild and extreme outliers
}
