% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFunctions.R
\name{Outliers}
\alias{Outliers}
\title{Outlier detection for numerical variables}
\usage{
Outliers(x)
}
\arguments{
\item{x}{= variable}
}
\value{
Returns a list with 4 objects showing mild and extreme outliers and the thresholds for each.
}
\description{
This function will return a list of 4 objects (1: a data frame with mild outliers and their indexes, 2: a data frame with extreme outliers and their indexes, 3: a vector with mild outlier thresholds, 4: a vector with extreme outlier thresholds)
}
