% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationfunctions.R
\name{rankCorrelations}
\alias{rankCorrelations}
\title{Ranks variable pair correlations}
\usage{
rankCorrelations(m, N = NULL, autoname = FALSE)
}
\arguments{
\item{m}{A correlation matrix}

\item{N}{Default \code{NULL} will return all correlations. Change to e.g. \code{N = 5}
to return only the 5 highest absolute correlations}

\item{autoname}{Automatically rename matrix columns and rows; default
is \code{FALSE}. Only recommended if \code{m} does not have names already.}
}
\value{
Dataframe with three columns; 1. variable 1, 2. variable 2 and
3. correlation
}
\description{
Takes in a correlation matrix and returns an ordered
dataframe of variable pairs. The order is from the absolute highest
to lowest correlation.
}
