% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryFunctions.R
\name{Summary}
\alias{Summary}
\title{Summary function that takes in a data frame or vector of categorical or numerical varibales and provides summary statistics}
\usage{
Summary(object)
}
\arguments{
\item{object}{can be a data frame or vector}
}
\value{
Data frame with relevant summary statistics
}
\description{
This function will provide relevant summary statistics for inputted data frames or vectors of categorrical or numerical variables
}
