% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryFunctions.R
\name{SummaryStatsCat}
\alias{SummaryStatsCat}
\title{Summary statistics for categorical variables}
\usage{
SummaryStatsCat(x)
}
\arguments{
\item{x}{= variable}
}
\value{
Data frame with summary statistics (count and percentage) for each level of the categorical variable
}
\description{
This function will return summary statistics (counts, percentage, missing values) for a specified categorical variable
}
