% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryFunctions.R
\name{SummaryStatsNum}
\alias{SummaryStatsNum}
\title{Summary statistics for numerical variables}
\usage{
SummaryStatsNum(x)
}
\arguments{
\item{x}{= variable}
}
\value{
vector with summary statistics
}
\description{
This function will return summary statistics (missing values,
mean, SD, SE, median, min, max, range, skew and kurtosis) for a specified numerical variable
}
