% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationfunctions.R
\name{targetCorrelations}
\alias{targetCorrelations}
\title{Correlations between a target variable and all other variables in a data frame}
\usage{
targetCorrelations(df, target, use = "everything", method = c("pearson",
  "kendall", "spearman"))
}
\description{
This function will return the correlation between a specfied target
variable and all other variables in the same data frame. Uses base R \code{cor}
to calculate correlations.
}
