% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryFunctions.R
\name{VSummary}
\alias{VSummary}
\title{Summary function for both categorical and numerical variables}
\usage{
VSummary(x)
}
\arguments{
\item{x}{A variable of type categorical or numerical}
}
\value{
Data frame with relevant summary statistics
}
\description{
This function will return relevant summary statistics dependent on the data type of the variable inputted
}
