% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationfunctions.R
\name{autoDensityPlot}
\alias{autoDensityPlot}
\title{Create a density estimate of a target variable}
\usage{
autoDensityPlot(df, target, interactiveplot = FALSE, xlabel = NULL, ...)
}
\arguments{
\item{df}{Dataframe that contains the target variable}

\item{target}{The name of the target variable}

\item{interactiveplot}{If \code{FALSE}, the default, returns a \code{ggplot2}
visualisation of the density estimate of the target variable. If
\code{TRUE}, returns an interactive \code{plotly} visualisation of the histogram.}

\item{xlabel}{Provide a character to override the default label for the
x axis}

\item{...}{Other arguments passed onto \code{geom_density}, such as
\code{colour = "red"} or \code{size = 2}}
}
\value{
A density estimate visualisation of the target variable from the
provided data
}
\description{
Generates either a static \code{ggplot} or interactive
\code{plotly} density estimate visualisation of the chosen target variable using
\code{geom_density}.
}
\examples{
mtcars \%>\%
  autoDensityPlot(hp)

# Can customise plot using usual ggplot2 aesthetics
mtcars \%>\%
  autoDensityPlot(hp, fill = "blue", alpha = 0.5)

}
