% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierFunctions.R
\name{OutlierMean}
\alias{OutlierMean}
\title{Mean with and without mild and extreme outliers}
\usage{
<<<<<<< HEAD
OutlierMean(outliers, x)
}
\arguments{
\item{outliers, }{the list output from the OutlierValues function, x, the variable of interest}
=======
OutlierMean(x)
}
\arguments{
\item{x}{= variable}
>>>>>>> master
}
\value{
A data frame with the mean without any outliers removed, the mean with mild and extreme outliers removed
}
\description{
This function will compute the mean of x with and without the mild and extreme outliers
}
